+++ src/nRGBlib/drawCircle.c
@@ -1,11 +1,11 @@
 #include <nGEO.h>
 
 /// Dessine un cercle en couleur
-void drawCircle_(short x, short y, short radius, Color c)
+void drawCircle_(int16_t x, int16_t y, int16_t radius, Color c)
 {
-    short x2 = 0;
-    short y2 = radius; // On se place en haut du cercle
-    short m = 5 - 4 * radius; // Initialisation
+    int16_t x2 = 0;
+    int16_t y2 = radius; // On se place en haut du cercle
+    int32_t m = 5 - 4 * radius; // Initialisation
 
     while (x2 <= y2) // Tant qu'on est dans le second octant
     {
